Straw-Hats/src/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow>=10.0.0
google-generativeai>=0.8.0
protobuf>=5.26.0
supabase==2.5.1
python-dotenv==1.0.1
pydantic>=2.7.0
httpx==0.27.0
